pyproject.toml
[project]
name = "ios-app-icon-generator"
version = "0.1.0"
description = "AI-powered iOS app icon generator using Stable Diffusion"
authors = [{name = "Cody Bontecou"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "replicate>=0.7.0",
    "pillow>=10.0.0",
    "rembg[cpu]>=2.0.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic[email]>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "authlib>=1.3.0",
    "httpx>=0.27.0",
]

[project.scripts]
icon-gen = "icon_generator.cli:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
